+++ date-of-love - (intellij)/src/main/java/br/com/dateoflove/servlet/AprovarOrcamentoServlet.java
@@ -32,7 +32,7 @@ public class AprovarOrcamentoServlet extends HttpServlet {
             UsuarioDao usuarioDao = new UsuarioDao();
 
             Usuario usuario =  usuarioDao.buscarUsuarioPorId(idUsuario);
-            Orcamentos orcamentos = orcamentosDao.buscarOrcamentoPorId(idOrcamento);
+            Orcamentos orcamentos = orcamentosDao.buscarOrcamentoPorId((long) idOrcamento);
 
 
             Email email = new Email("","",usuario.getEmail(),"","","");
